chore(deps): Add Abseil as a dependency
This commit adds abseil-cpp as a new external dependency using
CPMAddPackage in the root CMakeLists.txt.

It pins to the May 2025 LTS release (20250512.1) and disables
the building of Abseil's own tests to keep configuration and
build times minimal. This library is being added in preparation
for its use in upcoming changes.

Change-Id: I1587e03f4052bc170d07025117746a42b34dc82c